requirements.txt
langchain==0.0.350
langchain-community==0.0.20
langchain-openai==0.0.5
openai>=1.12.0,<2.0
chromadb==0.4.22
streamlit==1.34.0
tiktoken==0.5.2
python-dotenv==1.0.1
colorama==0.4.6
python-dateutil==2.8.2
requests==2.31.0